src/services/environment/pyproject.toml
[project]
  dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anthropic==0.49.0",
    "anyio==4.9.0",
    "asyncio==3.4.3",
    "attrs==25.3.0",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "distro==1.9.0",
    "fastapi==0.115.12",
    "filelock==3.18.0",
    "frozenlist==1.6.0",
    "fsspec==2025.3.2",
    "graphviz==0.20.3",
    "gunicorn==23.0.0",
    "h11==0.14.0",
    "httpcore==1.0.8",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.30.2",
    "idna==3.10",
    "iniconfig==2.1.0",
    "jiter==0.9.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain-anthropic==0.3.12",
    "langchain-core==0.3.54",
    "langchain-deepseek==0.1.3",
    "langchain-mistralai==0.2.10",
    "langchain-openai==0.3.14",
    "langchain-together==0.3.0",
    "langsmith==0.3.32",
    "loguru==0.7.3",
    "multidict==6.4.3",
    "openai==1.75.0",
    "orjson==3.10.16",
    "packaging==24.2",
    "pluggy==1.5.0",
    "propcache==0.3.1",
    "pydantic==2.11.3",
    "pydantic-core==2.33.1",
    "pytest==8.3.5",
    "pytest-asyncio==0.26.0",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "ruff==0.11.6",
    "sniffio==1.3.1",
    "starlette==0.46.2",
    "tenacity==9.1.2",
    "tiktoken==0.9.0",
    "tokenizers==0.21.1",
    "tqdm==4.67.1",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "yarl==1.20.0",
    "zstandard==0.23.0",
  ]
  description = "Add your description here"
  name = "environment"
  readme = "README.md"
  requires-python = ">=3.12"
  version = "0.1.0"
